fix(fish): remove setting shell if not needed
This removes the check on /etc/shells if you're not going to use Ansible
to provision fish as a shell. I'll freely admit that this is for a niche
use case around my work laptop and not being able to `su` to whatever I
need to do.

DCO-1.1-Signed-off-by: Patrick Wagstrom <patrick@wagstrom.net>