feat(fisher): add support for fisher plugins
This adds support for installating fisher alongside fish and also
installing plugins. This is only done if the installed version of fish
is greater than or equal to fish 3.0.0. It's smart enough to not keep
trying to install fish plugins that are already installed. But, it's not
yet smart enough to check if fisher is installed before proceeding, so
that gets triggered every time.

DCO-1.1-Signed-off-by: Patrick Wagstrom <patrick@wagstrom.net>