fix(config.fish): load symlinked files
The original implementation saw config.fish read in all files from
~/.config/fish/fish.d, but only files. If something was symlinked, it
wouldn't be read in. This caused issues installing the Google Cloud SDK
on MacOS using Homebrew because the .fish file needed should be
symlinked.

This small change makes it identify both files and symlinks. If MacOS
actually shipped with an up to date version of find, and we could be
ensured that this would only be installed on reasonably up to date
systems, we could use the `-xtype` parameter to find. But it appears
that option is only available in GNU find.

DCO-1.1-Signed-off-by: Patrick Wagstrom <patrick@wagstrom.net>