fix(homebrew): Fix for m1/m2 macOS paths
On the m1 and m2 versions of macOS, homebrew installs into /opt/homebrew
and not /usr/local. This means that hardcoded /usr/local paths don't
work. This is a not-so-elegant way to discover the appropriate prefix by
capturing the stdout of `brew --prefix` and then hacking together a
variable that can be used later in the role.

DCO-1.1 Signed-off-by: Patrick Wagstrom <160672+pridkett@users.noreply.github.com>